HX-Vector-Database-Server/0.1-Project-Execution/0.1.2-HXP-Shared-Library/requirements.txt
# HANA-X Vector Database Shared Library Dependencies
# Core Python version: 3.12+

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
strawberry-graphql>=0.216.0
grpcio>=1.59.0
grpcio-tools>=1.59.0

# HTTP Client and Async Support
aiohttp>=3.9.0
httpx>=0.25.0
asyncio-mqtt>=0.13.0

# Data Processing and Validation
pydantic>=2.5.0
numpy>=1.25.0
pandas>=2.1.0

# Database and Caching
redis[hiredis]>=5.0.0
qdrant-client>=1.7.0

# Monitoring and Metrics
prometheus-client>=0.19.0
structlog>=23.2.0
psutil>=5.9.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Security and Authentication
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Optional GPU Support (if available)
# torch>=2.1.0
# transformers>=4.35.0

# Production WSGI/ASGI Server
gunicorn>=21.2.0

# Utility Libraries
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
